FIR/UAST: commonize anonymous UClass
