+++ plugins/uast-kotlin-base/src/org/jetbrains/uast/kotlin/declarations/KotlinUAnonymousClass.kt
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2010-2021 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+
+package org.jetbrains.uast.kotlin
+
+import com.intellij.psi.PsiAnonymousClass
+import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiFile
+import org.jetbrains.kotlin.asJava.classes.KtLightClass
+import org.jetbrains.kotlin.psi.KtClassOrObject
+import org.jetbrains.kotlin.psi.KtObjectDeclaration
+import org.jetbrains.uast.*
+
+class KotlinUAnonymousClass(
+    psi: PsiAnonymousClass,
+    givenParent: UElement?
+) : AbstractKotlinUClass(givenParent), UAnonymousClass, PsiAnonymousClass by psi {
+
+    override val psi: PsiAnonymousClass = unwrap<UAnonymousClass, PsiAnonymousClass>(psi)
+
+    override val javaPsi: PsiAnonymousClass = psi
+
+    override val sourcePsi: KtClassOrObject? = ktClass
+
+    override fun getOriginalElement(): PsiElement? = super<AbstractKotlinUClass>.getOriginalElement()
+
+    override fun getSuperClass(): UClass? = super<AbstractKotlinUClass>.getSuperClass()
+    override fun getFields(): Array<UField> = super<AbstractKotlinUClass>.getFields()
+    override fun getMethods(): Array<UMethod> = super<AbstractKotlinUClass>.getMethods()
+    override fun getInitializers(): Array<UClassInitializer> = super<AbstractKotlinUClass>.getInitializers()
+    override fun getInnerClasses(): Array<UClass> = super<AbstractKotlinUClass>.getInnerClasses()
+
+    override fun getContainingFile(): PsiFile = unwrapFakeFileForLightClass(psi.containingFile)
+
+    override val uastAnchor by lazy {
+        val ktClassOrObject = (psi.originalElement as? KtLightClass)?.kotlinOrigin as? KtObjectDeclaration ?: return@lazy null
+        KotlinUIdentifier(ktClassOrObject.getObjectKeyword(), this)
+    }
+}
